MFH: Synchronized bundled GD library with GD 2.0.17

+++ ext/gd/libgd/gd_gd2.c
@@ -540,7 +540,7 @@ gdImagePtr gdImageCreateFromGd2PartCtx (gdIOCtx * in, int srcx, int srcy, int w,
 							}
 						} else {
 							ch = gdGetC(in);
-							if (ch == EOF) {
+							if ((int)ch == EOF) {
 								ch = 0;
 							}
 						}
